refactor: move search tests to main
